init: add tsconfig, esling, prettier,
